pyproject.toml
[tool.poetry]
name = "b2broker-test-task"
version = "0.1.0"
description = ""
authors = ["Anton Zakhlebayeu <antonzakhlebayeu@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.2.4"
djangorestframework = "^3.16.0"
psycopg2-binary = "^2.9.10"
djangorestframework-jsonapi = "^7.1.0"
gunicorn = "^23.0.0"
watchdog = "^6.0.0"
load-dotenv = "^0.1.0"
django-filter = "^25.1"
flake8-pyproject = "^1.2.3"
django-redis = "^6.0.0"
drf-spectacular = "^0.28.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.4.1"
pytest-django = "^4.11.1"
pytest-cov = "^6.2.1"
flake8 = "^7.3.0"
black = "^25.1.0"
isort = "^6.0.1"
pre-commit = "^4.2.0"

[tool.black]
line-length = 100

[tool.isort]
profile = "black"
default_section = "THIRDPARTY"
balanced_wrapping = true
known_third_party = ["rest_framework"]
known_first_party = ["api","ledger","project","tests"]
known_django = ["django"]
line_length = 88
lines_after_imports = 2
lines_between_sections = 1
multi_line_output = 3
sections = ["FUTURE","STDLIB","DJANGO","THIRDPARTY","FIRSTPARTY","LOCALFOLDER"]
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true

[tool.flake8]
exclude = [
        ".git", "__pycache__",
        ".tox", "admin.py", "manage.py",
        "*.example.py",
        "**/__init__.py", "apps.py", "settings", "**/migrations/*.py", "libs",
        "**/tests",
        "doc/*", "build","dist"
      ]
ignore = [
        "D100",   # Missing docstring in public module
        "D101",   # Missing docstring in public class
        "D102",   # Missing docstring in public method
        "D103",   # Missing docstring in public function
        "D105",   # Missing docstring in magic method
        "D106",   # Missing docstring in public nested class
        "D106",   # Missing docstring in public nested class
        "D107",   # Missing docstring in __init__
        "D202",   # No blank lines allowed after function docstring
        "D205",   # 1 blank line required between summary line and description
        "D301",   # Use r""" if any backslashes in a docstring
        "D400",   # First line should end with a period
        "D413",   # Missing blank line after last section
        "DAR000", # Attribute' object has no attribute 'id'
        "DAR101", # Missing parameter(s) in Docstring
        "DAR201", # Missing "Returns" in Docstring
        "DAR301", # Missing "Yields" in Docstring
        "DAR401", # Missing exception(s) in Raises section
        "E722",   # Do not use bare 'except'
        "E731",   # Do not assign a lambda expression
        "I101",   # Missing parameter(s)
        "I201",   # Missing "Returns"
        "I401",   # Missing exception(s) in Raises section
        "S001",   # Error parsing type params
        "W503",   # Line break occurred after a binary operator (opposite of W504)
        "F403",
        "F401",
        ]
max-complexity = 18
max-line-length = 120

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
